refactor: extract auth file restriction presenter
